requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==10.4.0
sumy==0.11.0
googletrans==4.0.0-rc1
